requirements.txt
fastapi
uvicorn[standard]
numpy
Pillow
httpx
